fix(storage): cast MDBX PID for Linux

+++ crates/storage/db/src/implementation/mdbx/mod.rs
@@ -442,7 +442,8 @@ impl DatabaseEnv {
             retry: std::ffi::c_int,
         ) -> HandleSlowReadersReturnCode {
             if space > MAX_SAFE_READER_SPACE {
-                let message = if is_current_process(process_id) {
+                #[allow(clippy::unnecessary_cast)]
+                let message = if is_current_process(process_id as u32) {
                     "Current process has a long-lived database transaction that grows the database file."
                 } else {
                     "External process has a long-lived database transaction that grows the database file. \
